Reorder workshop and update OpenAI example
