+++ readme.md
@@ -7,10 +7,10 @@ Welcome to the home for all Fermyon workshops!
 - [0. Setup](./spin/00-setup.md)
 - [1. Getting started with Spin](./spin/01-getting-started.md)
 - [2. Building a JSON API with Spin](./spin/02-json-api.md)
-- [3. Running your front-end application with Spin](./spin/03-frontend.md)
-- [4. Persisting non-relational data with Spin's key/value store](./spin/04-spin-kv.md)
-- [5. Using external databases](./spin/05-external-db.md)
+- [3. Using Fermyon Serverless AI from a JSON API](./spin/03-spin-ai.md)
+- [4. Running your front-end application with Spin](./spin/04-frontend.md)
+- [5. Persisting non-relational data with Spin's key/value store](./spin/05-spin-kv.md)
 - [6. Deploying to Fermyon Cloud](./spin/06-deploy-fermyon-cloud.md)
-- [7. Enhaving JSON API with Fermyon Serverless AI](./spin/07-fermyon-ai.md)
+- [7. Using external databases](./spin/07-external-db.md)
 - [8. Running your Spin application with Kubernetes](./spin/08-kubernetes.md)
 - (_optionally_) - Let us know what you think about this workshop using this short [form](https://form.typeform.com/to/RK08OLSy).
