+++ spin/03-spin-ai.md
@@ -231,6 +231,5 @@ In this section you learned how to:
 
 ## Navigation
 
-- Go back to [06 - Storing data in an external database](06-external-db.md) if you still have questions on
-  previous section
-- Otherwise, proceed to [08 - Deploying a Spin Application on Kubernetes](08-kubernetes.md)
+- Go back to [02 - Building a Magic 8 Ball JSON API with Spin](02-json-api.md) if you still have questions on previous section
+- Otherwise, proceed to [04 - Magic 8 Ball Frontend](04-frontend.md)
