+++ spin/07-external-db.md
@@ -1,4 +1,4 @@
-# Storing data in an external database
+# Storing key/value data in an external database
 
 Instead of using Spin's built in SQLite key/value store, you could use an external key/value store such as Redis. This section is intended as a learning exercise for the user. The complete code can be [found here](https://github.com/fermyon/workshops/tree/main/spin/apps/06-external-db)
 
@@ -18,5 +18,5 @@ In this section you learned how to:
 
 ### Navigation
 
-- Go back to [05 - Storing data in the Spin KV database](05-spin-kv.md) if you still have questions on previous section
-- Otherwise, proceed to [07 - Using Fermyon Serverless AI](07-fermyon-ai.md)
+- Go back to [06 - Deploying to Fermyon Cloud](06-deploy-fermyon-cloud.md) if you still have questions on previous section
+- Otherwise, proceed to [08 - Deploying a Spin Application on Kubernetes](08-kubernetes.md)
