+++ spin/apps/03-spin-ai/.gitignore
