+++ spin/apps/03-spin-ai/magic-8-ball-rust/Cargo.toml
