+++ spin/apps/03-spin-ai/magic-8-ball-ts/package.json
