+++ spin/apps/03b-openai/README.md
