+++ spin/apps/03b-openai/webpack.config.js
